requirements.txt
alembic==0.8.7
click==6.6
eventlet==0.19.0
Flask==0.11.1
Flask-Migrate==1.8.1
Flask-Script==2.0.5
Flask-SocketIO==2.6.1
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
python-editor==1.0.1
python-engineio==0.9.2
python-socketio==1.4.4
six==1.10.0
Werkzeug==0.11.10
gunicorn==0.16.1
Pillow==3.4.2
Flask-SQLAlchemy==2.1
Flask-Login==0.3.2
SQLAlchemy==1.0.15